Set a maximum total time for retries.
Summary:
Sometimes we do not want to retry an operation if it exeeded a
maximum time. E.g. in some integration tests we wait for an event with
retries. However, the future returned times out without giving any
reason. This change introduces a maximum time which defaults to
infinity. Once it's exeeded be throw an exception with the latest error.

Test Plan: unit-test, integration-test

Reviewers: meichstedt, aquamatthias, jasongilanfarr

Subscribers: marathon-team, jenkins

Differential Revision: https://phabricator.mesosphere.com/D224